+++ packages/browser-crawler/src/internals/browser-crawler.ts
@@ -10,7 +10,6 @@ import {
     ProxyInfo,
     RequestQueue,
     Session,
-    throwOnBlockedRequest,
     validators,
     storage,
     resolveBaseUrl,
@@ -39,7 +38,7 @@ import { BrowserLaunchContext } from './browser-launcher';
 
 export interface BrowserCrawlingContext<
     Page extends CommonPage = CommonPage,
-    Response = Dictionary<any>,
+    Response = Dictionary,
     ProvidedController = BrowserController,
 > extends CrawlingContext {
     browserController: ProvidedController;
@@ -64,14 +63,8 @@ export type BrowserHook<
     GoToOptions extends Record<PropertyKey, any> | undefined = Dictionary
 > = (crawlingContext: Context, gotoOptions: GoToOptions) => Awaitable<void>;
 
-export type GotoFunction<
-    Context = BrowserCrawlingContext,
-    GoToOptions extends Record<PropertyKey, any> | undefined = Dictionary
-> = (context: Context, gotoOptions?: GoToOptions) => Awaitable<any>;
-
 export interface BrowserCrawlerOptions<
     Context extends BrowserCrawlingContext = BrowserCrawlingContext,
-    GoToOptions extends Record<PropertyKey, any> | undefined = Dictionary,
     InternalBrowserPoolOptions extends BrowserPoolOptions = BrowserPoolOptions,
     __BrowserPlugins extends BrowserPlugin[] = InferBrowserPluginArray<InternalBrowserPoolOptions['browserPlugins']>,
     __BrowserControllerReturn extends BrowserController = ReturnType<__BrowserPlugins[number]['createController']>,
@@ -171,11 +164,6 @@ export interface BrowserCrawlerOptions<
     handlePageFunction?: BrowserCrawlerHandleRequest<Context>;
 
     /**
-     * Navigation function for corresponding library. `page.goto(url)` is supported by both `playwright` and `puppeteer`.
-     */
-    gotoFunction?: GotoFunction<Context, GoToOptions>;
-
-    /**
      * A function to handle requests that failed more than `option.maxRequestRetries` times.
      *
      * The function receives the following object as an argument:
@@ -272,11 +260,6 @@ export interface BrowserCrawlerOptions<
     navigationTimeoutSecs?: number;
 
     /**
-     * @deprecated Use `navigationTimeoutSecs` instead
-     */
-    gotoTimeoutSecs?: number;
-
-    /**
      * If cookies should be persisted between sessions.
      * This can only be used when `useSessionPool` is set to `true`.
      */
@@ -343,8 +326,6 @@ export abstract class BrowserCrawler<
 
     protected userProvidedRequestHandler!: BrowserCrawlerHandleRequest<Context>;
     protected navigationTimeoutMillis: number;
-    protected gotoFunction?: GotoFunction<Context, GoToOptions>;
-    protected defaultGotoOptions: GoToOptions;
     protected preNavigationHooks: BrowserHook<Context>[];
     protected postNavigationHooks: BrowserHook<Context>[];
     protected persistCookiesPerSession: boolean;
@@ -353,9 +334,6 @@ export abstract class BrowserCrawler<
         ...BasicCrawler.optionsShape,
         handlePageFunction: ow.optional.function,
 
-        gotoFunction: ow.optional.function,
-
-        gotoTimeoutSecs: ow.optional.number.greaterThan(0),
         navigationTimeoutSecs: ow.optional.number.greaterThan(0),
         preNavigationHooks: ow.optional.array,
         postNavigationHooks: ow.optional.array,
@@ -371,13 +349,11 @@ export abstract class BrowserCrawler<
     /**
      * All `BrowserCrawler` parameters are passed via an options object.
      */
-    protected constructor(options: BrowserCrawlerOptions<Context, GoToOptions>) {
+    protected constructor(options: BrowserCrawlerOptions<Context>) {
         ow(options, 'BrowserCrawlerOptions', ow.object.exactShape(BrowserCrawler.optionsShape));
         const {
             navigationTimeoutSecs = 60,
             requestHandlerTimeoutSecs = 60,
-            gotoFunction, // deprecated
-            gotoTimeoutSecs, // deprecated
             persistCookiesPerSession,
             proxyConfiguration,
             launchContext,
@@ -423,20 +399,9 @@ export abstract class BrowserCrawler<
             throw new Error('You cannot use "persistCookiesPerSession" without "useSessionPool" set to true.');
         }
 
-        if (gotoTimeoutSecs) {
-            this.log.deprecated('Option "gotoTimeoutSecs" is deprecated. Use "navigationTimeoutSecs" instead.');
-        }
-
         this.launchContext = launchContext;
-        this.navigationTimeoutMillis = (gotoTimeoutSecs || navigationTimeoutSecs) * 1000;
-
-        this.gotoFunction = gotoFunction;
-        this.defaultGotoOptions = {
-            timeout: this.navigationTimeoutMillis,
-        } as unknown as GoToOptions;
-
+        this.navigationTimeoutMillis = navigationTimeoutSecs * 1000;
         this.proxyConfiguration = proxyConfiguration;
-
         this.preNavigationHooks = preNavigationHooks;
         this.postNavigationHooks = postNavigationHooks;
 
@@ -565,7 +530,7 @@ export abstract class BrowserCrawler<
     }
 
     protected async _handleNavigation(crawlingContext: Context) {
-        const gotoOptions = { ...this.defaultGotoOptions };
+        const gotoOptions = { timeout: this.navigationTimeoutMillis } as unknown as GoToOptions;
 
         const preNavigationHooksCookies = this._getCookieHeaderFromRequest(crawlingContext.request);
 
@@ -577,7 +542,7 @@ export abstract class BrowserCrawler<
         await this._applyCookies(crawlingContext, preNavigationHooksCookies, postNavigationHooksCookies);
 
         try {
-            crawlingContext.response = await this._navigationHandler(crawlingContext, gotoOptions);
+            crawlingContext.response = await this._navigationHandler(crawlingContext, gotoOptions) ?? undefined;
         } catch (error) {
             this._handleNavigationTimeout(crawlingContext, error as Error);
 
@@ -614,14 +579,7 @@ export abstract class BrowserCrawler<
         }
     }
 
-    protected async _navigationHandler(crawlingContext: Context, gotoOptions: GoToOptions) {
-        if (!this.gotoFunction) {
-            // @TODO: although it is optional in the validation,
-            //   because when you make automation library specific you can override this handler.
-            throw new Error('BrowserCrawler: You must specify a gotoFunction!');
-        }
-        return this.gotoFunction(crawlingContext, gotoOptions);
-    }
+    protected abstract _navigationHandler(crawlingContext: Context, gotoOptions: GoToOptions): Promise<Context['response'] | null | undefined>;
 
     /**
      * Should be overridden in case of different automation library that does not support this response API.
@@ -632,7 +590,7 @@ export abstract class BrowserCrawler<
 
         if (this.sessionPool && response && session) {
             if (typeof response === 'object' && typeof response.status === 'function') {
-                throwOnBlockedRequest(session, response.status());
+                this._throwOnBlockedRequest(session, response.status());
             } else {
                 this.log.debug('Got a malformed Browser response.', { request, response });
             }