+++ test/apify/crawlers/basic_browser_crawler.ts
@@ -1,9 +1,13 @@
-import { BrowserCrawler, PuppeteerCrawlingContext, PuppeteerCrawlerOptions } from '@crawlee/puppeteer';
+import { BrowserCrawler, PuppeteerCrawlingContext, PuppeteerCrawlerOptions, PuppeteerGoToOptions } from '@crawlee/puppeteer';
 import { PuppeteerPlugin } from '@crawlee/browser-pool';
-import { LaunchOptions } from 'puppeteer';
+import { HTTPResponse, LaunchOptions } from 'puppeteer';
 
 export class BrowserCrawlerTest extends BrowserCrawler<{ browserPlugins: [PuppeteerPlugin] }, LaunchOptions, PuppeteerCrawlingContext> {
     constructor(options: Partial<PuppeteerCrawlerOptions> = {}) {
         super(options as any);
+    }
+
+    protected _navigationHandler(ctx: PuppeteerCrawlingContext, gotoOptions: PuppeteerGoToOptions): Promise<HTTPResponse | null | undefined> {
+        return ctx.page.goto(ctx.request.url, gotoOptions);
     }
 }